requirements.txt
streamlit
transformers
faiss-cpu
numpy
sqlalchemy
pymysql